\name{pStringDist}
\alias{pStringDist}
\title{Parallel hamming distance}
\description{Calculates the hamming distance matrix between a set of DNA
  sequences provided as a ShortReadQ or DNAStringSet objects.
}
\usage{
  pStringDist(data, output_filename)
}
\arguments{
  \item{data}{a character vector or an 'XStringSet' object}
  \item{output_filename}{results will be stored here as binary data}
}
\details{
  This is an early but fully working prototype.
}
\author{
  University of Edinburgh SPRINT Team
  \email{sprint@ed.ac.uk}
  \url{www.r-sprint.org}
}
\keyword{utilities}
\keyword{interface}

